# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Payment Gateway API
  version: "1.1.0"
  description: >
    USDT payment gateway supporting BEP20 (BNB Smart Chain) and TRC20 (Tron).
    Merchant endpoints authenticate with API key + HMAC signature; dashboard
    endpoints authenticate with JWT. The target chain is chosen per-payment via
    the optional `network` field (defaults to BEP20 for backward compatibility);
    TRC20 must be enabled server-side. Amounts are always human USDT strings —
    the gateway handles each chain's decimals (BEP20=18, TRC20=6) internally.
servers:
  - url: https://gateway.example.com/api/v1
  - url: http://localhost:4000/api/v1
...
components:
  securitySchemes:
    # Two API key modes coexist. Which one applies is decided by the key itself
    # (api_keys.auth_mode), not by the caller — presenting an HMAC key without a
    # signature, or a bearer token with one, is rejected rather than downgraded.
    #
    # 1. HMAC-signed: ApiKey carries the PUBLIC id (pk_live_…), and Signature +
    #    Timestamp must accompany it. The secret never crosses the wire.
    #
    #    X-Timestamp is unix SECONDS. The signed string is the timestamp, a
    #    literal ".", then the exact raw request body:
    #        X-Signature = hex( HMAC_SHA256( apiSecret, "${X-Timestamp}.${rawBody}" ) )
    #    For a request with no body (any GET) the raw body is "", so the signed
    #    string is "${X-Timestamp}.". Serialize the body once and send those
    #    exact bytes — re-serializing changes them and invalidates the
    #    signature. Requests more than 300 s from server time are rejected as
    #    replays. Verified against backend/src/middleware/auth.ts.
    ApiKey:      { type: apiKey, in: header, name: X-Api-Key }
    Signature:   { type: apiKey, in: header, name: X-Signature }
    Timestamp:   { type: apiKey, in: header, name: X-Timestamp }
    # 2. Bearer: the X-Api-Key value IS the secret (ak_live_…), sent alone.
    #    Simpler to integrate, but the credential is on the wire every call, so
    #    such keys are never granted the payouts:write scope.
    ApiKeyBearer: { type: apiKey, in: header, name: X-Api-Key }
    BearerAuth:  { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
...
    Payment:
      type: object
      properties:
        paymentId: { type: string, example: pay_01HZY... }
        orderId:   { type: string, example: order_789 }
        amount:    { type: string, example: "50.00" }
        amountReceived: { type: string, example: "0" }
        currency:  { type: string, example: USDT }
        network:   { type: string, enum: [BEP20, TRC20, ERC20, BTC], example: BEP20 }
        asset:
          type: string
          description: >
            Ledger asset; always equal to `currency`. Both are returned because
            `currency` is the historical field name and `asset` is what the
            multi-asset endpoints key on.
          example: USDT
        address:   { type: string, description: "BEP20 0x… or TRC20 T… deposit address", example: "0xabc..." }
        qrCode:    { type: string, description: base64 data URI }
        status:    { type: string, enum: [waiting,confirming,confirmed,partial,failed,expired,swept] }
        confirmations: { type: integer }
        txHash:    { type: string, nullable: true }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        fiat:
          $ref: '#/components/schemas/FiatQuote'
    FiatQuote:
      type: object
      description: >
        The frozen fiat quote, present ONLY on a payment that was priced in fiat.
        Its absence means the payment was priced directly in crypto.

        The rate is captured when the payment is created and is never recomputed:
        a customer quoted 52.581765 USDT for ₹5,000 owes that amount whenever they
        pay, and it is what the QR encodes and what confirmation is measured
        against. Treat it as an audit record of a decision, not a live price.
      properties:
        currency: { type: string, example: INR, description: ISO-4217 code the merchant priced in }
        amount:   { type: string, example: "5000", description: What the merchant asked for, in that currency }
        rate:
          type: string
          example: "95.09"
          description: Price of one unit of `asset` in `currency`, at `lockedAt`.
        source:
          type: string
          nullable: true
          example: coingecko
          description: >
            Provenance, including whether the rate was served stale
            (`coingecko:stale`). A stale quote is legitimate — the gateway
            prefers a slightly old price to a failed payment — but it is never
            silent.
        lockedAt:
          type: string
          format: date-time
          nullable: true
          description: When the rate was FETCHED upstream, not when it was used.
    CreatePayment:
      type: object
      required: [orderId]
      description: >
        Supply EXACTLY ONE pricing basis: `amount` (in crypto) or `fiatAmount` +
        `fiatCurrency`. Supplying both is a 400 — two prices would mean two
        answers to what the customer owes.
      properties:
        amount:
          type: string
          example: "50.00"
          description: Amount in `asset`. Omit when pricing in fiat.
        fiatAmount:
          type: string
          example: "5000"
          description: >
            Price in `fiatCurrency`. Converted to `asset` once, at creation, and
            frozen onto the payment (see FiatQuote). Returns 503 if no rate can
            be obtained at all — retry, or price directly in crypto.
        fiatCurrency:
          type: string
          example: INR
          description: >
            ISO-4217 code. Must be one this gateway can quote — see GET /rates.
            Required with `fiatAmount`.
        orderId:     { type: string, example: order_789 }
        description: { type: string, example: "Order #789 — Pro plan" }
        network:
          type: string
          enum: [BEP20, TRC20, ERC20, BTC]
          default: BEP20
          description: >
            Chain to settle on. Omit for BEP20 (back-compat). TRC20 requires the
            gateway to have Tron enabled, else the request returns 400.
...
paths:
...
  /payment-links:
    get:
      tags: [Payment links]
      summary: List this merchant's hosted-checkout links
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      responses:
        "200": { description: OK }
    post:
      tags: [Payment links]
      summary: Create a shareable checkout link
      description: >
        Returns a `token`. The customer-facing URL is
        `<panel>/pay/<token>`. Omit `amount` for an open amount (tips,
        donations). Pinning `asset` also requires `network` — the same symbol on
        two chains is two different assets.
      security: [{ BearerAuth: [] }, { ApiKey: [], Signature: [], Timestamp: [] }, { ApiKeyBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:       { type: string, maxLength: 120 }
                description: { type: string, maxLength: 500 }
                amount:      { type: string, description: Omit for an open amount }
                asset:       { type: string }
                network:     { type: string, enum: [BEP20, TRC20, ERC20, BTC] }
                reusable:    { type: boolean, default: true }
                maxUses:     { type: integer, minimum: 1 }
                expiresAt:   { type: string, format: date-time }
      responses:
        "201": { description: Created }
        "400": { description: Invalid amount, or an asset this gateway cannot settle }
  /pay/{token}:
    get:
      tags: [Payment links]
      summary: "PUBLIC: resolve a checkout link (no authentication)"
      description: >
        Reachable by anyone holding the link URL — the token IS the capability.
        Returns the business name, amount and accepted assets, and NOTHING that
        identifies the merchant further: no id, email, wallet, balance, webhook
        or keys. `usable: false` with `unusableReason` means expired, disabled or
        used up; render that as a dead end rather than an error.
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
        "404": { description: Unknown link, or the merchant is not approved }
        "429": { description: Rate limited }
  /pay/{token}/payments:
    post:
      tags: [Payment links]
      summary: "PUBLIC: start a payment from a link"
      description: >
        Claims one use of the link under a row lock, so a single-use link cannot
        be claimed twice by concurrent opens. Anything the link pins (amount,
        asset, network) overrides the request body — a crafted request cannot
        reduce a fixed price or pay in an excluded asset.
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                asset:   { type: string, description: Only used if the link does not pin one }
                network: { type: string }
                amount:  { type: string, description: Required only for open-amount links }
      responses:
        "201": { description: Deposit address and QR }
        "400": { description: Link expired, disabled, used up, or no amount given }
        "429": { description: Rate limited }
  /pay/{token}/payments/{paymentId}:
    get:
      tags: [Payment links]
      summary: "PUBLIC: poll a payment's status"
      description: >
        Scoped by the link token on purpose: a leaked payment id alone reveals
        nothing, and payments created through the API (which have no link) are
        never readable here.
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
        - { in: path, name: paymentId, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
        "404": { description: Not a payment of this link }
  /pay/{token}/invoice:
    get:
      tags: [Invoices]
      summary: "PUBLIC: the invoice behind a pay link"
      description: >
        An invoice's pay link IS its capability, so whoever holds the URL is
        exactly who should be able to read the document. Returns 404 when the
        link is an ordinary checkout link rather than an invoice — the checkout
        page probes this for every link and renders line items only if it gets
        something back.

        The response carries the document and the merchant's business name, and
        nothing else about the merchant: no id, email, wallet, balance, webhook,
        or other invoices.
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  number:       { type: string }
                  merchantName: { type: string }
                  customerName: { type: string, nullable: true }
                  currency:     { type: string }
                  subtotal:     { type: string }
                  taxAmount:    { type: string }
                  total:        { type: string }
                  notes:        { type: string, nullable: true }
                  dueDate:      { type: string, format: date, nullable: true }
                  status:       { type: string, enum: [open, paid, void] }
                  overdue:      { type: boolean }
                  issuedAt:     { type: string, format: date-time, nullable: true }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        description: { type: string }
                        quantity:    { type: string }
                        unitPrice:   { type: string }
                        amount:      { type: string }
        "404": { description: This link has no invoice, or does not exist }
...
  /payments:
    post:
      tags: [Payments]
      summary: Create a payment (generates a fresh BEP20 or TRC20 deposit address + QR)
      security: [{ ApiKey: [], Signature: [], Timestamp: [] }]
      parameters:
        - in: header
          name: Idempotency-Key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePayment' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        "400": { description: "Unsupported/disabled network", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "401": { description: Bad signature, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    get:
      tags: [Payments]
      summary: List payments (paginated)
      security: [{ ApiKey: [], Signature: [], Timestamp: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: page,   schema: { type: integer, default: 1 } }
        - { in: query, name: limit,  schema: { type: integer, default: 25 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Payment' } }
                  page: { type: integer }
                  total: { type: integer }
  /payments/{id}:
    get:
      tags: [Payments]
      summary: Retrieve a payment
      security: [{ ApiKey: [], Signature: [], Timestamp: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Payment' } } } }
        "404": { description: Not found }
...